requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
langchain
langchain_google_genai
google-generativeai
python-dotenv
